good progress: metaclass methods almost work - but not quite
Test with:

    cargo run foo/impl/test_transpile.foo --use=foo/lib -- --class

problem in #testRuntimeClass2:

    class MyMetaclass {}
        is Class
    end

produces an instance represented by an empty array, not a "proper class",
and uses foo_mark_instance (or rather foo_mark_array). So the class
instance we create is not actually an instance of that, even though we say
it is. Which breaks GC. Since Class is a concrete class, other metaclasses
are (1) instances of it, and (2) inherit it as their superclass.

This means that MyMetaclass should have the the same layout as Class.

This also means that metaclass argument is redundant, and better expressed
as:

    class MyMetaclass {}
        is Class
        method ping
            "pong"!
    end

    let MyClass = MyMetaclass name: "MyClass"
                              layout: Layout empty
                              methods: [].
     MyClass ping --> #pong

problem in testRuntimeClass3:

    We should use method_arity instead of selector arity, but doing
    so corrupts the heap and it's not clear why.